+++ test/libsolidity/syntaxTests/pragma/experimental_solidity_out_of_order_2.sol
@@ -1,0 +1,13 @@
+function f() pure returns (uint)
+{
+    return 1;
+}
+
+pragma experimental solidity;
+
+struct A
+{
+    uint256 x;
+}
+// ----
+// ParserError 8185: (83-89): Experimental pragma "solidity" can only be set at the beginning of the source unit.